SPRING 2021 Home Announcements Assignments Zoom Discussions Grades Friday Open Lab Repl.it Github Syllabus Virtual Textbooks Unit 2 - Mario Project 2 Unit 2 - Mario Project 2 Due: Fri Feb 5, 2021 11:59pmDue: Fri Feb 5, 2021 11:59pm Unlimited Attempts ungraded/100 Points –/100 Points Calculated by: Most Recent Available Available complete Upload in-progress Submit incomplete Not Graded Yet incomplete Mario 2
In Mario 1, you implemented a function to print a "pyramid" to the developer console. Today, you will put that pyramid where it truly belongs: on your web page! To render the pyramid to the page, you will use Javascript to programmatically insert it into the DOM.

Setup
Open the Mario Pyramid repl you created in Mario Project 1
We will be upgrading the project to utilize the DOM and dynamically update the web page!
You will need to replace the code in your index.html file with the following:
<!DOCTYPE html>
<html>
<head>
    <style>
        #pyramid {
            font-family: "Courier New", monospace;
        }
    </style>
</head>
<body>
    <h1>Mario 2</h1>
    <h3>
        Behold this wonderful pyramid:
    </h3>

    <div id="pyramid">
        <!-- Our pyramid will go inside here -->

        <div id="construction">
            <p>Uh oh! the pyramid is still under construction.</p>
            <p>Check back again soon... our developers are hard at wrk as we speak!</p>
        </div>
    </div>

    <script src="mario.js"></script>
</body>
</html>
Take a Look
View the index.html page in a browser window, you will see that the text on the page has changed since last time. It now says:

Behold this wonderful pyramid!

The user is no longer instructed to open the Dev Tools console because the pyramid is supposed to show up directly on the page. But, of course, instead of a pyramid, we see the same notice as last time:

Uh oh! the pyramid is still under construction.

Now, if you do happen to open up the Console, you will see a pyramid show up down there.

Note also that instead of spaces (" "), you shouldshould update your function to use ("."). We'll talk about that further down. For now, just think of them as grains swirling in the desert wind.

Before getting started, quickly look at index.html. Observe that inside the <body> is a <div> whose id attribute is equal to "pyramid". We have placed this element here as a "container" which is sitting there waiting to be "injected" with the actual contents of the pyramid.

Your Task
Complete the TODO at the bottom of the loop inside drawPyramid, so that each row gets inserted into the DOM as a child element of the container

 <div id="pyramid">.

// TODO 1
        // create an element whose inner text is rowStr,
        // and insert it as a child of the container <div id="pyramid">

In the scope of this loop, you will need a variable called rowStr, containing a string for the current row. You just need to add that string to the DOM.

Create a <p> element and set its text to be the string, and then insert that paragraph element as a child of the container <div> on mario.html.

 - You can build and add elements, check out this example: https://www.w3schools.com/js/js_htmldom_nodes.asp

Once you have the pyramid showing up, delete the "Under Construction" notice from the container <div>.
// TODO 2
// delete the "under construction" element, ie the <div id="notice">
Note: What's up with the periods instead of spaces? Well, just try switching back now and see what happens. As you have seen before, HTML tends to compress whitespace. This is often very convenient, but in this particular case, it messes up the alignment of our pyramid. A more reasonable thing to do, rather than use a period character, would be to use an HTML-encoded space character. We will use this technique in the next studio.

Here is what the page will look like with a height of 10.

mario2doneoutput.JPG

Committing Your Changes
Click the version control tab in repl.it.

Then commit, along with a descriptive message.

$ git commit -m "render pyramid to the DOM instead of the console"